+++ src/pages/link-preview.js
@@ -36,7 +36,7 @@ import {
 
 import ArrowLink from 'components/patterns/ArrowLink'
 import CaptionBase from 'components/patterns/Caption/Caption'
-import { CUSTOMERS } from 'components/patterns/CustomerStory'
+import { CUSTOMERS, customerPath } from 'components/patterns/CustomerStory'
 import Faq from 'components/patterns/Faq/Faq'
 import Features from 'components/patterns/Features/Features'
 import FetchProvider from 'components/patterns/FetchProvider'
@@ -1733,7 +1733,7 @@ const CustomerStories = () => {
         {stories.map(({ slug, name, blurb, icon }) => (
           <CustomerStoryCard
             key={slug}
-            href={`/use-cases/${slug}`}
+            href={customerPath(slug)}
             data-event-location='LinkPreview'
             data-event-name={`Customer Story ${name}`}
           >